Add aggregate macro and #changed? helper method
Adds `.aggregate` macro to define which attribute is the aggregate object
that the command is acting on. Calling the macro defines the aggregate
name and adds an attr_accessor.

```
class UserCommand
  prepend ActiveModel::Command

  aggregate :user
end
```

Adds `#aggregate` instance method which returns the aggregate if it has
been defined with the macro.

Adds `#changed?` helper method which checks if an attribute was given
and is different from the aggregate's value for that attribute. This
method is dependent on the aggregate having been defined with the macro.

```
class UserCommand
  prepend ActiveModel::Command

  aggregate :user

  attr_accessor :name

  def call
    if changed?(:name)
      user.name = name
    end
    user
  end
end
```

For array attributes, it is assumed that order does not matter. In cases
where order does matter, `true` can be passed as a second argument to
compare them strictly.

```
class UserCommand
  prepend ActiveModel::Command

  aggregate :user

  attr_accessor :sorted_tags

  def call
    if changed?(:sorted_tags, true)
      user.sorted_tags = sorted_tags
    end
    user
  end
end
```