Add subject macro and #changed? helper method (#8)
Adds `.subject` macro to define which attribute is the subject
that the command is acting on. Calling the macro defines the subject
name and adds an attr_accessor.

```
class UserCommand
  prepend ActiveModel::Command

  subject :user
end
```

Adds `#subject` instance method which returns the subject if it has
been defined with the macro.

Adds `#changed?` helper method which checks if an attribute was given
and is different from the subject's value for that attribute. This
method is dependent on the subject having been defined with the macro.

```
class UserCommand
  prepend ActiveModel::Command

  subject :user

  attr_accessor :name

  def call
    if changed?(:name)
      user.name = name
    end
    user
  end
end
```

For array attributes, it is assumed that order does not matter. In cases
where order does matter, `true` can be passed as a second argument to
compare them strictly.

```
class UserCommand
  prepend ActiveModel::Command

  subject :user

  attr_accessor :sorted_tags

  def call
    if changed?(:sorted_tags, true)
      user.sorted_tags = sorted_tags
    end
    user
  end
end
```